Channels/web/requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
colorama==0.4.3
coverage==5.2
dnspython==1.16.0
email-validator==1.1.1
eventlet==0.25.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
flask-marshmallow==0.13.0
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
greenlet==0.4.16
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
monotonic==1.5
more-itertools==8.4.0
mypy==0.781
mypy-extensions==0.4.3
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
python-engineio==3.13.0
python-socketio==4.6.0
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.17
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.1
